#32 入力Viewの実装分割
