+++ calc/Views/ExprInputView.swift
@@ -12,37 +12,22 @@ struct ExprInputView: View {
     @Binding var viewModel: CalcModel
     @FocusState.Binding var textEditorFocused: Bool
 
+    enum InputViewType: Int {
+        case full = 1
+        case bySentence =  2
+    }
+
+    
     var body: some View {
         VStack(alignment: .leading, spacing: 0) {
             Text("式を入力")
                 .modifier(LabelTextModifier())
-            TextEditor(text: $editText)
-                .focused(self.$textEditorFocused)
-                .keyboardType(.numbersAndPunctuation)
-                .scrollContentBackground(Visibility.hidden)
-                .modifier(MultiLineTextModifier())
-                .frame(maxWidth: .infinity,
-                       minHeight: 50.0,
-                       maxHeight: /*@START_MENU_TOKEN@*/.infinity/*@END_MENU_TOKEN@*/,
-                       alignment: .topLeading)
-                .onAppear() {
-                    // UIApplication.shared.closeKeyboard()
-                }.onChange(of: self.editText) { oldText, newText in
-                    viewModel.expr = newText
-                    viewModel.calc()
-                }
-        }
-        .onTapGesture {
-            self.textEditorFocused = false
+            AnyView(buildInputView())
         }
     }
-    private func buildExprVariables() -> [ExprVariable] {
-        guard let symbolTable = viewModel.symbolTable else {
-            return []
-        }
-        return symbolTable.asArray.map { symbolElement in
-            ExprVariable(name: symbolElement.name, value: symbolElement.value)
-        }
+    
+    private func buildInputView() -> any View {
+        FullExprInputView(editText: editText, viewModel: $viewModel, textEditorFocused: $textEditorFocused)
     }
 }
